+++ packages/widget_toolkit/lib/src/lib_language_picker/theme/language_picker_theme.dart
@@ -153,11 +153,6 @@ class _$LanguagePickerTheme {
     WidgetToolkitDesignSystem.light().icons.checkIcon,
   ];
 
-  static List<SvgFile> leftInfoCircleIcon = [
-    WidgetToolkitDesignSystem.dark().icons.infoCircleIcon,
-    WidgetToolkitDesignSystem.light().icons.infoCircleIcon,
-  ];
-
   /// endregion
 }