+++ packages/widget_toolkit/test/lib_language_picker/view/language_picker_golden_test.dart
@@ -7,16 +7,8 @@ void main() {
     generateDeviceBuilder(
         widget: languagePickerSuccessPageFactory(
           showError: false,
-          hasLeftIcon: false,
           isLoading: false,
         ),
         scenario: Scenario(name: 'language_picker_success')),
-    generateDeviceBuilder(
-        widget: languagePickerSuccessPageFactory(
-          showError: false,
-          hasLeftIcon: true,
-          isLoading: false,
-        ),
-        scenario: Scenario(name: 'language_picker_has_left_icon')),
   ]);
 }